Wrote HOC with-genres and with-title-click
